requirements.txt
Django==4.0.4
djangorestframework==3.13.1
gunicorn==20.1.0
clickhouse-driver==0.2.5
django-clickhouse-backend==1.0.1
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.9
